dashboard/requirements.txt
dash
plotly
pandas
asyncio
websockets
flask
azure-data-tables
dotenv
geopandas
shapely
dash_bootstrap_components